requirements.txt
cryptography==3.4.7
distlib==0.3.3
fastapi==0.66.0
passlib==1.7.4
pydantic==1.8.2
python-jose==3.3.0
SQLAlchemy==1.4.21
SQLAlchemy-Utils==1.4.21
uvicorn==0.14.0

